FU specific stuns, weak and strong
