+++ stats/effects/stun/fustunstrong.statuseffect
@@ -1,0 +1,15 @@
+{
+  "name" : "fustunweak",
+  "blockingStat" : "stunImmunity",
+  "effectConfig" : {"strong":true},
+  "defaultDuration" : 1.5,
+
+  "scripts" : [
+    "fustun.lua"
+  ],
+
+  "animationConfig" : "stun.animation",
+
+  "label" : "Stun",
+  "icon" : "/interface/statuses/slow.png"
+}
