+++ stats/effects/stun/fustunweak.statuseffect
@@ -1,0 +1,15 @@
+{
+  "name" : "fustunweak",
+  "blockingStat" : "stunImmunity",
+  "effectConfig" : {"strong":false},
+  "defaultDuration" : 1.5,
+
+  "scripts" : [
+    "fustun.lua"
+  ],
+
+  "animationConfig" : "stun.animation",
+
+  "label" : "Stun",
+  "icon" : "/interface/statuses/slow.png"
+}
